aether mind beta

+++ xRLAgents/utils/Augmentations.py
@@ -10,7 +10,7 @@ def aug_random_apply(x, p, aug_func):
 
 #uniform aditional noise
 def aug_noise(x, k = 0.2): 
-    pointwise_noise = k*(2.0*torch.rand(x.shape, device=x.device) - 1.0)
+    pointwise_noise = k*(2.0*torch.rand(x.shape, device=x.device, dtype=x.dtype) - 1.0)
     return x + pointwise_noise  
 
 # random black mask
@@ -18,10 +18,10 @@ def aug_mask(x, p = 0.75, gw = 16, gh = 16):
     up_h = x.shape[2]//gh
     up_w = x.shape[3]//gw 
 
-    mask = torch.rand((x.shape[0], x.shape[1], gh, gw), device = x.device)
+    mask = torch.rand((x.shape[0], x.shape[1], gh, gw), device = x.device, dtype=x.dtype)
     
     mask = torch.nn.functional.interpolate(mask, scale_factor = (up_h, up_w), mode="bicubic")
-    mask = (mask > (1.0 - p)).float().detach()
+    mask = (mask > (1.0 - p)).to(x.dtype).detach()
 
     return mask*x
